requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.10.4
pyyaml==6.0.2
